Dev nores 02c

+++ autorino/common/step_cls.py
@@ -89,8 +89,14 @@ class StepGnss:
             The session information for the step. If not provided, a dummy session is created.
         options : dict, optional
             The options for the step. If not provided, an empty options dictionary is created.
-        metadata : dict, optional
-            The metadata for the step. If not provided, a dummy metadata is created.
+        metadata : str or list, optional
+            The metadata to be included in the converted RINEX files. Possible inputs are:
+             * list of string (sitelog file paths),
+             * single string (single sitelog file path)
+             * single string (directory containing the sitelogs)
+             * list of MetaData objects
+             * single MetaData object.
+             Defaults to None.
         """
         self._init_epoch_range(epoch_range)
         self._init_site(site)
@@ -1157,10 +1163,14 @@ class StepGnss:
         list
             The list of paths of the invalidated files.
         """
-        med = self.table["size_out"].median(skipna=True)
-        valid_bool = threshold * med < self.table["size_out"]
-        self.table.loc[:, "ok_out"] = valid_bool
-        invalid_local_files_list = list(self.table.loc[valid_bool, "fpath_out"])
+
+        if not self.table["size_out"].isna().all():
+            med = self.table["size_out"].median(skipna=True)
+            valid_bool = threshold * med < self.table["size_out"]
+            self.table.loc[:, "ok_out"] = valid_bool
+            invalid_local_files_list = list(self.table.loc[valid_bool, "fpath_out"])
+        else:
+            invalid_local_files_list = []
 
         return invalid_local_files_list
 
@@ -1355,7 +1365,7 @@ class StepGnss:
         # TEMP RINEX Files
         tmp_rnx_files_new = []
         for f in self.tmp_rnx_files:
-            if os.path.isfile(f):
+            if f and os.path.isfile(f):
                 logger.debug("remove tmp converted RINEX file: %s", f)
                 os.remove(f)
             else:
@@ -1376,14 +1386,14 @@ class StepGnss:
 
             is_original = self.table["fpath_ori"].isin([f]).any()
 
-            if os.path.isfile(f) and is_original:
+            if f and os.path.isfile(f) and is_original:
                 logger.warning(
                     "uncompressed file is also an original one, we keep it for security: %s",
                     f,
                 )
                 tmp_decmp_files_new.append(f)
                 continue
-            elif os.path.isfile(f) and not is_original:
+            elif f and os.path.isfile(f) and not is_original:
                 logger.debug("remove tmp decompress RINEX file: %s", f)
                 os.remove(f)
             else:
@@ -1862,7 +1872,7 @@ class StepGnss:
 
             self.write_in_table_log(self.table.loc[irow])
         else:
-            ### update table if things go wrong
+            # ++ update table if things go wrong
             self.table.loc[irow, "ok_out"] = False
             self.write_in_table_log(self.table.loc[irow])
             # raise e
